Reasons to use double
Hello Henry,

Here are some counter-arguments in favour of double. I absolutely think BigDecimal has it's place, but I think using it as the default, outside of money-handling, is a premature optimisation. 

Having written this up though, I reckon this page may be a bit niche for a book that focuses on general-purpose core coding issues. It might be worth dropping -- or moving to another book!

 - Daniel